refactor: inline fullscreen state transitions
