+++ lib/features/scenes/presentation/providers/player_view_mode.dart
@@ -1,1 +1,3 @@
 enum PlayerViewMode { inline, fullscreen, tiktok }
+
+enum FullscreenPhase { inline, entering, fullscreen, exiting, tiktok }
